Reset _selected _pressed and _dragging rows when filtered
We get _selected, _pressed and _dragging values (Dialogs::Row*) from IndexedList::all() method. That method returns _pFiltered variable, but after this we filter the list of dialogs and reset the _pFiltered value to new list and new values (we delete the previous list items). And as a result we get undefined behavior because we have wandering pointer (pointer that does not point to existing variable).